Pass down absolute paths to DaemonicParserState to fix cross cell include_def invalidation
Summary:
Add a base path parameter to watchman events, so that each event knows which cell
the path that's normally passed in is relative to. Additionally, use this to
pass in the full path to invalidatePath, instead of using the old behavior,
which simply prepends the cell root onto each path.

This should allow changes to build file defs, included from another cell, to
properly signal invalidation for the cell in which it was used.

We should probably do more work to audit how relative and absolute paths are
used in these caches, as other similar issues may still be present, but that's
outside the scope of this diff (I hope).

Test Plan: CI, added a test.

Reviewed By: mzlee

fbshipit-source-id: a175dd2