Apple Library: add flag to indicate that a target is modular.
Summary:
This adds an opt-in flag that can be added to Apple library targets when used as frameworks so they're built as modular. The option is backwards compatible (off by default) and future proof as it should eventually be supported by Buck itself.

To make a particular target modular, add the following to the definition in your `BUCK` file:

    modular = True

In order for Xcode to build a modular framework:

- Headers need to be part of a build phase so that they're recognised by Xcode. This enables recognition of an umbrella header required for Swift & Obj-C interop.
- VFS overlays are required to enable compilation of modular frameworks with Swift & Obj-C interop. In order to generate those overlays, both `USE_HEADERMAP` and `CLANG_ENABLE_MODULES` must be turned on. VFS overlays are passed to the compiler via `-ivfsoverlay`.
- Buck-generated header maps break the VFS overlays for public headers which ends up triggering a failure in recognising said headers as modular (as they're expected to be found in `$BUILT_PRODUCTS_DIR/Module.framework/Versions/A/Headers`). To work around this behaviour, public headers for modular frameworks are not part of header maps or symlink trees.

Test Plan: Unit tests and internal FB app testing.

Reviewed By: dinhviethoa

fbshipit-source-id: 33a3266