Don't allow empty jars from dummy R.java
Summary:
We were seeing (incorrectly) empty jars showing up in cache from
machines with full disks.  This should detect and prevent that from
happening again.

There might be some rare cases where an empty R.java jar is correct.  If
that becomes a problem, we can detect it and disable this check.

Test Plan:
CI

Will try to catch logs in out-of-disk situations in production.

Reviewed By: jkeljo

fbshipit-source-id: 3ccd9ed