Allow cxx_library to be precise about which header dependencies are re-exported
Summary:
Previously, all dependencies of a cxx_library are automatically re-exported
when compiling binaries and libraries that depend on it. This doesn't capture
the common case of implementing an API without exposing dependencies which are
only implementation detail. This causes libraries and individual compile rules
at the top of a large dependency graph to needlessly include transitive
dependencies which are not exposed in the public API of libraries, leading to
huge include search paths, unnecessary rule key misses, and additional
processing and memory usage in buck when computing input based rule keys.

This diff introduces a backward compatible way of specifying that only
libraries listed in `exported_deps` are re-exported for sake of header
dependencies. It adds a `reexport_all_header_dependencies` field which controls
whether all `deps` are exported, or only `exported_deps`. The field is default
true, preserving the old behavior of re-exporting all deps. This allows the
option to be turned on on a library-by-library basis for incremental migration
purposes.

Test Plan: Added an integration test.

Reviewed By: Coneko

fbshipit-source-id: 1056fbb