fix watchman globs in non-root cells
Summary:
Wathcman glob in non-root cells was broken due to the stats filter not taking
into account of the cell path. It used to stat basePath + filePath, which is
relative to the current directory (i.e. root cell). I change this to use the
absolute path instead.

Test Plan:
Enable watchman globbing in another cell, run `buck targets blah//blah:blah`
and see that the globbed sources appear, and no warnings are emitted.

Reviewed By: mzlee

fbshipit-source-id: d9fe4a4