Make getting the hg root interruptible.
Summary:
Computing the hg root requires shelling out which makes it
interruptible, but we currently swallow the interruption
request and make it fail. This has two consequences: the
interrupt will not stop Buck, and the computation will fail as
if the hg root call failed, even though it might have succeeded
had it not been interrupted.

Test Plan: CI

Reviewed By: aiked

fbshipit-source-id: 299305c