Add support for BuildRuleEvent.Suspended and Resumed in DistBuildSlaveEventBusListener.
Summary:
Right now the DistBuild SuperConsole shows
```
SERVER 1)=> WORKING ON 2563 JOBS ...
```
which is not true because we didn't factor in the Suspended and Resumed BuildRuleEvents.
This diff fixes this statistic to a more sane value of
```
SERVER 1)=> WORKING ON 40 JOBS ...
```
which is rightly the number of threads in the slave.

Test Plan: Unit test. Also tried running a distributed build with dev version of buck.

Reviewed By: alisdair04

fbshipit-source-id: bf61d17