Use "origin" build ID fields to chain builds that cached/re-cached artifacts
Summary:
This changes the semantics of the new "origin" build ID field.  Previously,
this always referred to the build ID that built the artifact.  However,
multiple builds may end up fetching and re-caching the same artifact, which
meant we didn't have a way to trace these hops the artifact took.

This diff now uses the "origin" build ID to store the build ID of the previous
build that *cached* the artifact, which could be either due to building it
locally or re-caching it off a cache fetch.  This means we can use the "origin"
build ID field as a sort of linked list, which iterates through all the builds
which handled the artifact, which should be strictly more useful than the
previousl semantics.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 2bbb8a6