Fix haskell linker error message
Summary: Was trying to use Buck with Haskell for a side project and getting an error about not having the compiler flag set in my .buckconfig despite having set it. Dug through the source code and realized that it was actually missing the linker flag but that the error message was incorrect. :-/

Test Plan: bakedcaketest

Reviewed By: niteria

fbshipit-source-id: c21d716