CSharpLibrary fixes
Summary:
The previous version of test runs only when `csc` is available on PATH. It is not the case by default. Test was updates to run with VS 2015 by default (see https://msdn.microsoft.com/en-us/library/1700bbwd.aspx).
References should not be escaped since the whole argument is escaped (It seems that automatic escaping of arguments was introduced a long time ago, and we did not run CsharpLibraryIntegrationTest).

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: db89058