Fixed: Ensure entries spooled directly to jar are written in sorted order
Summary:
[Resubmitted after previous revert. The issue was that the changes to `JavaInMemoryFileObject` caused
it to write empty files if a `getJavaFileForOutput` was called without then opening a stream and writing to
it. The standard implementation of `Filer.getResource` will call this method any time a `Location` is
specified that is typically used for output.]

We were depending on the compiler to write these in a deterministic order. And it does, but
there are no guarantees about the behavior of annotation processors or other plugins.

This change makes it such that files written by the compiler are written to the jar in sorted order,
followed by any other files that were needed (also in sorted order). That's still not identical
to `spool_mode = intermediate_to_disk`, but it is at least deterministic even when the compiler or
its plugins are not.

Test Plan: `JavaInMemoryFileObjectTest`, `JavaInMemoryFileManagerTest`, and `DefaultJavaLibraryIntegrationTest`

Reviewed By: asp2insp

fbshipit-source-id: 0c8fa6e