Cache generated BuildRules for xcode project generation
Summary:
Previously, we constructed all BuildRules in a subgraph repeatedly when we
requested a BuildRuleResolver for xcode project generation. We do this as the
requested BuildRules are often sparse, so generating the entire action graph
can be wasteful. However, also wasteful is the re-generation build rules of new
action graphs.

Here I introduce a hack that's a sort of middle-ground from always
re-generating subgraphs and generating a single dense action graph of all
nodes. LazyActionGraph will ensure the subgraph from a node is fully
instantiated in a bottom-up manner, in the same way that the previous code did,
but share this build rule resolver across all invocations.

Note that this is all necessary because BuildRuleResolver is currently not
threadsafe, once it becomes thread safe, we can simply pass around the
BuildRuleResolver itself without this hack.

Reviewed By: dinhviethoa

fbshipit-source-id: e3a4a6b