Introduce and integrate CalculateAbiFromSource
Summary:
`CalculateAbiFromSource` will compute an ABI jar for a `JavaLibrary` rule
using just the sources (plus dependencies, for now).

With this commit it is now possible to generate ABIs from source, with some
caveats:
- only for rules that don't run annotation processors
- the ABI rule will have all the main rule's dependencies available to it
- ABIs generated this way are not yet identical to ABIs generated from class files, so there are likely bugs still

Test Plan:
Enable `abi_generation_mode = source_with_deps` in `.buckconfig.local`,
verify in traces and `buck-out` that many rules now have `#source-abi` flavors
instead of `#class-abi` ones.

This change itself is pretty straightforward (`CalculateAbiFromSource` is
a very simple class), so I'm not really sure what automated tests are worth
writing, but I'm open to ideas.

As for the integration, the following commit will introduce a validation
mode that will generate both source- and class-based ABIs and diff them.
The results of running that on a few nontrivial builds will be used to
fill holes in `StubJarTest` and shore up this feature.

Reviewed By: dreiss

fbshipit-source-id: c3a72a6