Replaced FileHashCache with a prefix tree.
Summary:
This changes the file hash cache to a prefix tree instead of a hash map. It reduces times for invalidation, although no reliable benchmark is available.

The CacheBenchmark.java benchmark is supposed to be measuring it but the results are a bit odd and not reflecting the speedup introduced (which is otherwise noticeable via plain timestamps)

Test Plan: Tests and Sandcastle

Reviewed By: aiked, illicitonion, marcinkosiba

fbshipit-source-id: 48b4fd3