dbuck: always run the buck that's next to it
Summary:
`dbuck` used to run whatever `buck` happened to be on the path, which
was fine if you ran `dbuck` from the path.

This commit makes it run the `buck` that's in the same directory with it,
so that you can run `bin/dbuck` to debug your locally-built `buck`, even
if your path points to a different `buck`.

Test Plan: `bin/dbuck build buck`

Reviewed By: illicitonion

fbshipit-source-id: e3506f6