Remove metadata from DB in `DefaultOnDiskBuildInfo.deleteExistingMetadata`
Summary:
When a build rule fails we delete its on-disk metadata; we need to
delete it from the BuildInfoStore as well.  This addresses cases where we remove
filesystem metadata (e.g., RECORDED_PATHS), but not DB metadata (e.g, RULE_KEY),
which leaves us in an inconsistent state that can't be recovered from without a
clean (because all future builds will try to read RECORDED_PATHS and throw when
it can't be found).

Test Plan:
I don't have a solid repro to add as a test case, unfortunately. I
had (locally) enabled some logging that showed this method was called on a
problematic target, and subsequent builds failed in the expected manner.

Reviewed By: marcinkosiba

fbshipit-source-id: a44dbb4