Start reworking StubJarTest using JarDiffer
Summary:
The test logic in `StubJarTest` is getting pretty complex, and it's hard to
tell what exactly is being tested in any given place. This commit begins to
rework the testing logic as a textual compare of the ASM dump of the stub jar.
That makes the differences very clear, and also allows the test author to comment
on the expected dump to call out important features if desired.

This commit is a direct migration of many of the simpler tests. Following commits
will migrate the rest.

Test Plan:
These are tests ;-)

More seriously, the way each of these tests was created by taking the stub
output of previously-validated tests. ASM's stub printer includes all the
details of the class file, so that gives me confidence that we're continuing
to test all of the same stuff.

Reviewed By: asp2insp

fbshipit-source-id: 347ebd8