Migrate StubJarTest tests that compare stub jars to the new model
Summary: TSIA

Test Plan:
These are tests ;-)

More seriously, the way each of these tests was created by taking the stub
output of previously-validated tests. ASM's stub printer includes all the
details of the class file, so that gives me confidence that we're continuing
to test all of the same stuff.

Reviewed By: asp2insp

fbshipit-source-id: 5aa2641