Migrate StubJarTest tests that work with compiler-generated elements to the new model
Summary:
There are a few tests that are trying to test stubbing behavior with compiler-generated
elements (synthetics and bridges). We get more confidence in such tests if we can see
what the compiler generated to know that the test is actually exercising what we want it
to be.

Test Plan:
These are tests ;-)

More seriously, the way each of these tests was created by taking the stub
output of previously-validated tests. ASM's stub printer includes all the
details of the class file, so that gives me confidence that we're continuing
to test all of the same stuff.

Reviewed By: asp2insp

fbshipit-source-id: 1d06087