StubJarTest: Delete Manifest test
Summary:
The stub jar manifest won't always be generated by the `StubJar` class. It's better to test
this at a higher level. The existing test in `CalculateAbiFromSourceIntegrationTest` and the
new one added to `CalculateAbiFromClassesStepTest` in this commit are a better way to
make sure this is happening.

Test Plan: See description

Reviewed By: asp2insp

fbshipit-source-id: 3b0d1a3