Correct access flags of private inner classes
Summary:
Inner classes were bolted on to the class file format after a couple of
releases had gone by. As such, the class access flags can only express
`public` and default accessibility, and the accessibility of an inner class
is actually found on its `InnerClass` attribute.

Empirically, however, it appears that `javac` does change the class access
flags for an inner class based on its true visibility, mapping `protected`
to `public` and `private` to default. We were getting that wrong in one case
in `ClassVisitorDriverForElement`.

Test Plan: New unit test

Reviewed By: asp2insp

fbshipit-source-id: f77e060