Experimental `buck project --view` command
Summary:
This is a follow-up to an internal Ruby script which prototyped the Project View functionality. It's a pretty literal Java port of the Ruby prototype. I want to keep a tight correspondence between the prototype and this experimental `buck` command: I'll address any functional and/or Buck-style comments in follow-up diffs.

Instead of creating an IntelliJ module for each Buck module, `buck project --view` simply creates a folder full of symlinks, and a single IntelliJ module that represents the new 'project view'. The idea is that IntelliJ is known to work very well with very large numbers of files in a modest number of modules, but not so well with a large number of modules. Using a project view should let you edit without stalls.

So far, Cmd+O and Cmd+B seem to work just fine;  I only see one class of any red identifiers that I don't see in a 'normal' `buck project` project.

There is one known issue and one possible one:
 1. The command creates a single Android module for the whole project. One thing this means is that trying to use Android classes (like `android.text.TextUtils.isEmpty`) in a non-Android Buck module will //not// show as invalid. There are ways to fix this issue ... but my suspicion is that it is not going to be a very large problem for most people.

 * I still totally haven't looked into how change notifications will work in a project view! My //expectation// is that IJ's native watcher will do the right thing, but this is something that needs testing ....

There are two issues in this diff that are fixed later in the stack:
 1. The command creates a symlink for each and every resource file. If you add a resource file (or get one //via// a VCS pull) **//and//** you need to see it in IntelliJ, you'll need to rerun `buck project --view`  to regenerate the view. I'll add view update logic in a diff on top of this one; for now, though, you need to regen. (Fixed in the  `updateView` diff; note that you still have to rerun `buck project --view` //but// it is now safe to do so 'on top of' an existing view - it will delete stale links, create new ones as necessary, and leave valid ones in place.)
 2. The command will only link to a single copy of any 'multi-value' resource. When you have, say, `drawable/foo.png` and `drawable-hdpi/foo.png`, it will link  `drawable/foo.png` under `/res`. But! When you have, say,  `drawable-hdpi/foo.png` and  `drawable-mdpi/foo.png` but no `drawable/foo.png`, you will end up with a `/res/drawable-...` resource. This should have no functional impact, but is a bit messy: I'll fix it in a follow-up diff. (Fixed in the `dedash` diff.)

Test Plan:
The command takes at least 2 arguments: directory to create, and target(s). For example, `buck project --view ~/project_view project_target`.

As a prototype, the command does almost no error checking!

 * The target can be an `.buckconfig` alias. You can pass multiple targets (so that you can, say, jump back and forth between your code and your JUnit tests) but there 'should' be no more than one `app` level target. (It will just pick the 1st one, if it there are more than one.)
 * The directory should not exist! I have not tested this scenario, but "weird things" //may// happen if you try to create a project view in an existing directory. (This restriction was removed in the `updateView` diff.)

Run the command, then do File / Open ... (or just click the icon on the toolbar). Select the directory the script command.

Reviewed By: styurin

fbshipit-source-id: a861a69