Clean output directories if metadata_storage is inconsistent
Summary:
When a user changes the build.metadata_storage format, we need to ensure that
stale metadata is not left behind, or future builds could use incorrect
metadata, leading to broken builds.  Unfortunately, using the filesystem to
store metadata makes it hard to reliably find metadata for all targets, so the
safest thing to do is remove the entire output directory.

Test Plan:
- Updated integration test
- Manual testing switching from filesystem -> rocksdb -> filesystem -> rocksdb

Reviewed By: asp2insp

fbshipit-source-id: fda8822