(IjProject) Use labels to add path with generated sources
Summary:
We use dependencies as a marker to detect if generated source paths needs to be added to the project. Instead, we can use labels to do that same. This should be more efficient because the number of labels is usually much lower than the number of dependencies.

Also, keeping the old logic in place to make transitioning easier. It will be removed after we migrate our repo to the new model.

Test Plan: Unit tests

Reviewed By: asp2insp

fbshipit-source-id: d187738