Consolidate calculation of cache error rate with CacheRateStatsKeeper.
Summary:
We noticed that cache error rate for remote builds was not being calculated quite the same as we do
in local builds. This diff fixes that by transmitting a serializable form of CacheRateStatsKeeper
from the distributed BuildSlave to the client rather than passing each metric on its own. We then
piggy-back on CacheRateStatsUpdateEvent to get us the same miss and error rates as we would have
calculated in a local build.

Test Plan: CI

Reviewed By: alisdair04

fbshipit-source-id: 253f908