Cache jar content listing immediately after writing
Summary:
When computing the dep-file rule key for a target, we need to enumerate
the contents of all of its dependencies. The previous implementation was
resulting in opening the same jars multiple times during a build. Now we
read the jar directory shortly after writing each jar to disk (either when
built locally or fetched from cache) and keep it on the build rule, thus
guaranteeing that each jar is only read once for this purpose.

Future work could look at obtaining this listing *while* writing the jar
rather than shortly thereafter.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 25e337d