Better hashCode caching in Pair
Summary:
The existing cache involved two objects (one of which was a `WeakReference`) just
to cache an `int`. That's an 8-byte pointer, at least 40 bytes worth of objects, and some
GC overhead, plus we GC so often the hashCodes were getting recomputed all the time anyway.

A regular old `int` and `volatile boolean` is much more memory and GC-efficient, plus it
allows for double-checked locking.

Test Plan: CI

Reviewed By: asp2insp, dreiss

fbshipit-source-id: 9efbe4c