Exit fast from LoggingBuildListener.handleFallback()
Summary:
This gets a lot of events and spends a not insignificant amount of time
checking if it's a type it cares about. Change it to check if logging is
even enabled before the more expensice class check.

Test Plan: do noop build of large app, see it spend less time in LoggingBuildListener

Reviewed By: Coneko

fbshipit-source-id: 2a7e0a2