Eradicate PrintWriter usage in Buck
Summary:
`PrintWriter` swallows `IOException`s, which may come as a surprise to
those used to working with the rest of Java's IO APIs. We've found one
cases where this led to artifact corruption in low-disk-space scenarios,
so this commit takes the pre-emptive step of moving to our own version
that throws.

Test Plan: Code inspection

Reviewed By: dreiss

fbshipit-source-id: 89a2908