Fix buckd on Debian/Ubuntu by using bash
The `buckd` script uses `bash` features (as opposed to generic bourne
shell `sh` features) like the `function` keyword and the `BASH_SOURCE`
array.

This is a problem if `/bin/sh` isn't just a symlink to `bash`, eg in
Debian/Ubuntu it points to `dash`.

Use `/bin/bash` explicitly (just as is already done in the main `buck`
script).