Merge linkables across the whole app to ensure the dependencies are consistent and then sort into apk modules
Summary: we hit an issue when modules are viewed independently, the constinuent ordering calculations don't look into merge constituents in other modules.  This updates the code to do everything globally and then verify the module relationships.

Test Plan: built a modular build with a native library in an app module that also depends on a linkable in the root module

Reviewed By: JamieEi

fbshipit-source-id: 7342b8d