Return an Immutable instead of a Future from CachingBuildEngine.build
Summary:
To make sure we completed uploading all artifacts to the cache we currently
wait on the IO executor to shutdown. I'd like to make the wait explicit.
For that I need to return both the 'result' future and 'upload' future from the
build method. This is just a refactoring that introduces the new class and makes
the BuildEngine return it.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 73fa14f