pass `--release` flag to worker for libraries
Summary:
If a `js_library` is built with the `#release` flavor, the worker now gets a `--release` flag passed in.

Background:
Until recently, the library step in the worker only concatenated transformation result. Now, we also generate code for assets at this stage.
To generate code correctly, we need information about whether we are doing a release build or not.

Test Plan: Adapt integration tests, run all tests.

Reviewed By: jeanlauliac

fbshipit-source-id: b9718df