Improve error handling of malformed cache manifests
Summary:
Previously, a malformed cache manifest could crash the entire build.
Now, we log it as an error (which goes to the console), then just
pretend that there was no manifest.  The newly-constructed manifest
should overwrite the old one in cache (if writes to cache are enabled).

Test Plan:
Ran a build that was previously crashing due to a bad manifest.  Saw
these two errors in the console, then the build completed successfully
after building the rule locally.

fbshipit-source-id: 0b6ae4f