Remove code that falls back to building locally after failing to unzip a cache artifact
Summary:
This code was written as a crutch against inadequate failure logging.
Now failure logging is adequate, so we no longer need it.  I've never
seen this code be useful in the last few years, and it was causing
problems recently by hiding a disk-full situation (which should lead to
rapid failures).

Test Plan: Forced an exception.  Saw it reported to the console.

Reviewed By: marcinkosiba, jkeljo

fbshipit-source-id: a35dbd9