Add SQLiteBuildInfoStore
Summary:
Might as well get this code reviewed.  It's fast, and it's robust
against Ctrl-C, buck kill, and pkill.  It's got native code which is kind of a
bummer, but (a) I've figured out why RocksDB was crashing, and I'll fix that,
(b) SQLite is supremely battle-tested so any problems are clearly due to my
usage, and (c) I'll write a native-crash detecting wrapper before this gets
turned on anywhere.

I promise to remove the unused BuildInfoStores once I'm done experimenting, but it's super-helpful to have them in-tree now.

Test Plan: local stress testing

Reviewed By: marcinkosiba

fbshipit-source-id: e574113