Make the lifetime of BuildInfoStoreManager longer than the CommandThreadManager
Summary:
BuildRule futures can outlive the CachingBuildEngine, for various
reasons (perhaps they ignore InterruptedException, for example).  This creates a
problem when the CachingBuildEngine closes, and in turn closes its
BuildInfoStores, since those long-lived futures can later attempt to write to
the now-closed Store.

Fix this situation by making the BuildInfoStoreManager a long-lived object;
placing it in CommandParams allows it to be easily propagated to the creation
sites of CachingBuildEngines, and it naturally outlives the
CommandThreadManager, which will reap all of the BuildRule futures when it shuts
down.

Test Plan: RocksDBBuildInfoStore no longer segfaults on Ctrl-C

Reviewed By: marcinkosiba

fbshipit-source-id: 46617a9