Move IJ-specific part of ProjectCommand to IjProjectCommandHelper
Summary:
`ProjectCommand` supports two IDEs at this moment (InteliiJ and XCode) and a lot of code is shared between two project generators. Changing shared logic is difficult because these changes need to be tested in both generators.

This is the first part of refactoring that splits logic in ProjectCommand into independent pieces.

Some IntelliJ-specific logic is moved to `IjProjectCommandHelper`. There are no changes in logic, only refactoring and plumbing.

Note: `IjProjectCommandHelper` cannot depend on `CommandRunnerParams` so it accepts all necessary objects from `CommandRunnerParams` as arguments. `BuckBuildRunner` is added to provide a way to run a build from the helper.

Test Plan: Unit tests.

Reviewed By: dinhviethoa

fbshipit-source-id: 25ef0fc