(IjProject) Move parsing logic to IjProjectCommandHelper
Summary:
`ProjectCommand` supports two IDEs at this moment (InteliiJ and XCode) and a lot of code is shared between two project generators. Changing shared logic is difficult because these changes need to be tested in both generators.

This is the second part of refactoring that splits logic in ProjectCommand into independent pieces.

The rest of IntelliJ-specific logic is moved to `IjProjectCommandHelper`. Also, duplicated logic responsible for parsing targets, moved it to `IjProjectCommandHelper` and simplified by removing logic that handles XCode. All in one change to avoid having a dependency in IntelliJ package on XCode classes.

Test Plan: Unit tests

Reviewed By: dinhviethoa

fbshipit-source-id: 6d38fbb