Move XCode project generation logic to XCodeProjectCommandHelper
Summary:
`ProjectCommand` supports two IDEs at this moment (InteliiJ and XCode) and a lot of code is shared between two project generators. Changing shared logic is difficult because these changes need to be tested in both generators.

This is the last part of refactoring that splits logic in ProjectCommand into independent pieces.

XCode logic is moved to XCodeProjectCommandHelper.

Test Plan: Unit tests

Reviewed By: dinhviethoa

fbshipit-source-id: 8f05d1f