Make bash variables ignore special characters in for javac.sh
In travis I am seeing these errors:

```
====ERROR LOGS====
[2017-05-08 05:33:49.554][warn ][tid:553][com.facebook.buck.apple.AppleConfig] Could not execute xcode-select, continuing without developer dir.
[2017-05-08 05:33:49.756][warn ][tid:552][com.facebook.buck.event.listener.LoggingBuildListener] BUILD FAILED: //java/com/example:example failed with exit code 2:
/tmp/junit-temp-path8236425419225122826/javac.sh
stderr: + ARGS_FILE=buck-out/gen/java/com/example/__example__srcs
+ shift
/tmp/junit-temp-path8236425419225122826/javac.sh: 16: /tmp/junit-temp-path8236425419225122826/javac.sh: Syntax error: "(" unexpected
```

There is only one instance of `(` in the file...and AFAICT it is valid bash. So let's assume that one of the args from the
args file has `(` in it and quote the variable.