Add a busy handler, use WAL journaling
Summary:
We want to allow concurrent access to the database, but we hold a
connection for the duration of the command.  But if we re-enable WAL journaling
and add a busy handler, concurrent access works fine.  I perf tested WAL,
and it's about as fast as OFF as far as I can tell.

Test Plan:
perf tests
simultaneous buck build and buck {run,test}
pkill/buck kill during a build and observe data is still written

Reviewed By: dreiss

fbshipit-source-id: e23b9d4