Refactor NdkCxxPlatforms to get cxxflags from buckconfig
Summary:
The flags passed into to the compiler during android builds is not configurable, so thread in the AndroidBuckConfig to NdkCxxPlatforms so we can ask for flags during Platform creation.  Hard coding compiler flags and preprocessor flags makes it difficult to do project-wide compiler flag upgrades.

However, to not break expected behavior, add `extra_c*flags` instead of replacing all of the default flags. And there are a lot of default flags

Test Plan: Run tests and add new unit tests

Reviewed By: marcinkosiba

fbshipit-source-id: 144b26e