Improve efficiency of manifest caching loads and stores
Summary: Instead of constructing a multimap of `Path -> [SourcePath]` for every reachable file, precompute the entries that are needed and build a Multimap of only needed entries. As the number of files actually used for a compilation is usually much fewer than the number of files that are visible to a rule, this dramatically reduces manifest caching load and store times for rules that have many transitive dependencies.

Test Plan:
The effects of this (at least for writes) can be profiled fairly easily.

I disabled http cache to only use the dir cache, and used a cache-breaker preprocessor flag to force a rebuild of only this library.

```
buck build //path/to:lib#static,iphonesimulator-x86_64 --cache.mode=dir
```

For a target containing 50k visible header files in the universe, and 732 source files to compile, `Manifest.addEntry` time taken reduced from 34 seconds 5.5 seconds:

| Method                                    | File          | Diff (ms) | Old (ms) | New (ms) |
| com.facebook.buck.rules.Manifest.addEntry | Manifest.java | -28626    | 34129    | 5503     |

Reviewed By: andrewjcg

fbshipit-source-id: 8a50997