Do not throw an exception if DirCache fails to list cache files
Summary:
Sometimes DirCache fails to list all files inside it's cache folder. This results to a crash
which happens from time to time during clean up operation. This should not kill the build as this is
a cache problem and we tend to keep caches to be non-build-breaking.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 53699c7