Support com.apple.developer.icloud-services entitlement for iOS apps
Summary:
Adding a support for iCloud entitlements in iOS app entitlements in buck. iCloud entitlement in an app entitlements file (*.plist file) looks like this:

  <key>com.apple.developer.icloud-services</key>
  <array>
    <string>CloudDocuments</string>
  </array>

In a provisioning profile iCloud entitlement looks like this:

  <key>com.apple.developer.icloud-services</key>
  <string>*</string>

When buck matches provisioning profiles to app entitlements, it looks for either direct match or requires that an array of entitlements is a subset of an array of provisioning profile entitlements which is not the case here since provisioning profile entitlement is a wildcard character `*`.

This diff adds support for `com.apple.developer.icloud-services` entitlement by adding a new rule: if provisioning profile has a wildcard `*` for this key then consider this a match regardless of app entitlements.

Test Plan:
1. Add iCloud entitlement to iOS app provisioning profile:
```
  <key>com.apple.developer.icloud-services</key>
  <array>
    <string>CloudDocuments</string>
  </array>
```

2. Build it with buck:
```
  buck build adsmanager#iphoneos-arm64
```

Reviewed By: ryu2

fbshipit-source-id: 0fc2ce1