Improve memory usage of MergeAssetsStep
Summary:
D4743133 introduced a change which put a lot of pressure on
the Java heap. While this didn't cause OOMs, it was often the
victim.

The root cause was Buck reading the entire contents of each binary
resource into heap memory in order to:

1) Calculate their CRC32 hash code and length, then
2) Write them to a ZIP file

Thankfully, we don't need the entire contents of these files in
heap memory to do these operations.

This diff avoids the memory overhead by:

1) Re-using the existing hash code and length when copying resources from one ZIP file to another
2) When merging resources, using Guava `ByteSource` to allow us to first calculate the CRC32 hash code and length, then copy the resource into the ZIP file without reading the entire thing into memory

Test Plan: `ant java-test -Dtest.class=AndroidBinaryIntegrationTest`

fbshipit-source-id: 95bda83