Upload native crashes to logview
Summary:
If we're going to use native code for the DB (and I think SQLite is a
winner), we should have some logging in place for native crashes.  This dumps
logs when a JVM error log file is produced.

Test Plan:
wrote a little JNA helper to let me call `strlen(NULL)`, and verified
that it crashes the JVM and that logs are dumped.

Reviewed By: mjpieters

fbshipit-source-id: e990702