(IjProject) Reduce the number of folders in IJ module
Summary:
The current algorithm of simplifying source folders in a given module is very conservative and results in a lot of duplicated declarations.

For example, it can create the following structure:

```
      <sourceFolder url="file://$MODULE_DIR$/module" isTestSource="false" packagePrefix="com.example.module" />
      <sourceFolder url="file://$MODULE_DIR$/module/submodule" isTestSource="false" packagePrefix="com.example.module.submodule" />
```

In this case the second line is unnecessary because it's package name matches the parent's package name + folder name.

Merging folders is using bottom-up recursive processing of a file tree constructed from these folders. At each node its children folders are analyzed and some of them are merged into a parent.

When there is a folder in a parent tree node only children that match the type of the parent folder are merged into the parent.

In case when a parent folder doesn't exist the algorithm tries to create a new parent node in the following order (creation succeeds if there is at least one child of the given type):
- source folder without package information (code in non-Java languages)
- source folder with package information (code in Java-languages)
- test folder without package information
- test folder with package information

Note that the algorithm can merge children excluded folders into a parent excluded folder only if the parent folder exists and there are no children of other types. This is because IntelliJ doesn't handle non-excluded folders under excluded folders very well (it hides these folders in Project View if hiding excluded folders is enabled).

Test Plan: Unit tests and local testing.

Reviewed By: JonShemitz

fbshipit-source-id: a298b58