(IjProject) Change how merged folder is created
Summary:
The current algorithm to create a new parent results in a very unbalanced structure when a single `SourceFolder` in children can result in a parent of this type and the rest of `TestFolder`s will be kept in the iml.

To address this problem the algorithm is changed to create a parent of the type that have the biggest number of children. In the mentioned example it would create `TestFolder` instead of `SourceFolder`.

Test Plan: Unit tests and local testing.

Reviewed By: asp2insp

fbshipit-source-id: 2b40d8e