Prefix tree is now thread safe.
Summary:
The cache needed to be put in thread safety and after thinking a bit how to do it, I opted for synchronizing all public methods. The impact on this should be quite minimal and should serve the case. After all the main issue that the prefix tree is addressing is the overpopulation of the cache that results in buck hanging (that is, cutting the spikes in https://fburl.com/aa0wxewt), so even at a minimal performance cost, this should do the trick and not be too expensive.

I left some comments explaining the thready safety concerns and the solution in place.

Test Plan: CI and local test

Reviewed By: illicitonion

fbshipit-source-id: 678087e