Improve performance of classpath computation
Summary:
None of the callsites of this function expect a sorted set, so
a normal set will do just as well. This eliminates an expensive call to
`ImmutableSortedSet.copyOf` at each node which is queried for its classpath.

For posterity, at this time this function is called from four places, all of which have their own sorting implemented when necessary:
1) `GraphEnhancementQueryEnvironment#getFirstOrderClasspath` which returns a Stream already (so clients will sort as necessary)
2) `UnsortedAndroidResourceDeps#createFrom` where it is used to seed a graph traversal (the results of which are marked as unordered anyway)
3) `JavaLibraryClasspathProvider#getTransitiveClasspathDeps` which returns an unordered set
4) `AndroidTransitiveDependencyGraph#findManifestFiles` which returns an unordered set

Test Plan:
Benchmark action graph, check time spent in query has gone down.
F67006415
(old and new are reversed in that comparison, so this represents ~2500ms win)

Reviewed By: illicitonion

fbshipit-source-id: b495466